Split out the post-request methods from the base class.
