+++ src/lib/list.ts
@@ -1,5 +1,5 @@
 import { Types } from "../mapper";
-import { Base } from "../utils";
+import { Base, Request } from "../utils";
 import { Web } from "./web";
 
 /*********************************************************************************************************************************/
@@ -19,7 +19,7 @@ class _List extends Base {
         this.targetInfo.endpoint = "web/lists/getByTitle('" + listName + "')";
 
         // Add the methods
-        this.addMethods(this, { __metadata: { type: "list" } });
+        Request.addMethods(this, { __metadata: { type: "list" } });
     }
 
     // Method to get the list by the entity name.
