+++ src/utils/index.ts
@@ -4,5 +4,6 @@ export * from "./dependencies";
 export * from "./methodInfo";
 export * from "./oData";
 export * from "./promise";
+export * from "./request";
 export * from "./targetInfo";
 export * from "./xhrRequest";
